restructured some files and folders
